Fix OOM crashes during screenshot comparison
- Recycle captured bitmap in createBitmapFromActivity() after saving to
  disk, eliminating ~9.6 MB leaked per test invocation
- Switch IntBuffer allocation from heap (IntBuffer.allocate) to native
  memory (ByteBuffer.allocateDirect) to bypass JVM heap constraints
- Add overflow guard for large capacity * INTEGER_BYTES computations
- Wrap ParallelPixelProcessor analyze() and transform() in try/finally
  to guarantee buffer cleanup on exceptions
- Recycle diff bitmap in HighContrastDiff.generate() after saving
- Update ImageBufferTest to reflect direct buffer allocation behavior

+++ Library/src/main/java/dev/testify/core/exception/LowMemoryException.kt
@@ -50,7 +50,7 @@ import dev.testify.core.formatDeviceString
 @Suppress("ktlint:standard:argument-list-wrapping")
 class LowMemoryException(
     targetContext: Context,
-    requestedAllocation: Int,
+    requestedAllocation: Long,
     memoryInfo: String,
     cause: OutOfMemoryError
 ) : TestifyException(
